pyproject.toml
[project]
name = "remedia"
version = "0.3.0"
description = "Duplicate detection for media files (images, videos, more)"
authors = [
  { name = "Christian Seip", email = "christian@seip.io" }
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
  "pillow>=10.0.0",
  "imagehash>=4.3",
  "tqdm>=4.66",
  "torch>=2.0",
  "open-clip-torch>=2.20"
]

[project.scripts]
remedia = "remedia.main:run_cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
